Make AddComponent into an extension method
