--- Unity/Assets/Utilities/Il2Cpp/Il2CppUtils.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: fb8897f63d2a4c469956bc0bedfccca2
-timeCreated: 1738523439
